requirements.txt
python-dotenv
requests
pywin32
